+++ examples/example/vue3/src/examples/grid/01/App.vue
@@ -4,9 +4,14 @@ import { Col } from '@expcat/tigercat-vue/Col'
 </script>
 
 <template>
-  <Row :gutter="[16, 16]">
-    <Col v-for="item in 3" :key="item" :span="8">
-      <div class="rounded bg-blue-600 p-4 text-center text-white">col-8</div>
-    </Col>
-  </Row>
+  <div class="rounded border border-[var(--tiger-border)] p-3">
+    <Row :gutter="[16, 16]">
+      <Col v-for="item in 3" :key="item" :span="8">
+        <div
+          class="rounded bg-[var(--tiger-primary)] p-4 text-center text-[var(--tiger-primary-foreground,#fff)]">
+          col-8
+        </div>
+      </Col>
+    </Row>
+  </div>
 </template>